refactor: move ExtContext code to shared activation
This moves the creation of the ExtContext object in to the
shared activation method, and additionally moves some other
minor things to ensure this works.
