+++ packages/toolkit/src/shared/extensionGlobals.ts
@@ -123,6 +123,7 @@ interface ToolkitGlobals {
     readonly context: ExtensionContext
     // TODO: make the rest of these readonly (or delete them)
     outputChannel: OutputChannel
+    invokeOutputChannel: OutputChannel
     loginManager: LoginManager
     awsContextCommands: AwsContextCommands
     awsContext: AwsContext
